update interactive.py
When trying to detect format of the training data file in the _write_nlu_to_file function, there is a conflict between the opening of the file in writing mode and the call to the "_guess_format" fonction from rasa_nlu/rasa_nlu/training_data/loading.py.

Indeed, after the opening, the file is everytime detected as markdown by the "_guess_format" func because the "js = json.loads(content)" loads an empty file resulting in wrong detection (if the file is a json). 

Quick fix to avoid that.