keyword intent classifier now takes keywords from training data
This commit replaces the previous simpler keyword_intent_classier which
simply treated all words found in training data as keywords. With the
new implementation keywords can be added to an intent by adding them to
examples with the syntax `[intent[keyword]]`, where `keyword` is the
keyword that should cause any utterance containing it to be classified
as the intent specified in `intent`.
If `intent` is empty the keyword will be assigned to the intent under
which the keyword is marked.
Keyword matching is done by checking whether the keyword string is
contained in a message string, case is ignored.
Things to note about the current implementation:
With the current implementation keywords can contain whitespaces, if
this is the case only the exact string (module case) will result in the
message being classified based on the keyword. E.g. if the keyword is
"some keyword" then neither "some" nor "keyword" will trigger the
keyword_intent_classier.
Classification in general proceeds by having the classifier set the
`intent` property of the `Message` object. This means that a classifier
will overwrite any Classification that has already been done. Therefore
the keyword_intent_classier will only have an effect if it is listed
after the main classifier in the config.yml.