Fix TypeError with SklearnPolicy
When training `SklearnPolicy` the method to retrieve the
model failed when non-featurizer parameters were passed to
`.train`, e.g. failing with:

    TypeError: model_architecture() got an unexpected keyword
               argument 'dump_stories'

While `model_architecture` filtered the params, at some point
the `**kwargs` parameter was lost. This issue is fixed
by reintroducing the parameter and updating the training
parameters with the `kwargs` passed to `train`.