pin networkx
networkx 2.4 breaks visualization (i.e. `rasa interactive` as well as the visualization on rasa x -- viz just doesn't appear). pinning to 2.3 fixes this. Tested on rasa x by extending the rasa x image to install 2.3 and visualizations re-appeared. this should be in the next patch so that we can create a rasa x patch that pulls from this rasa version.