Include story source filename in core output for failed stories
Include story source filename in the story name in the failed stories output to help find the failed story more easily (see #3419).
Passed the source filename starting from the `StoryFileReader` to a `StoryStep`.
Besides the story block names the source filename is included in the tracker events which are used for outputting the failed stories.

Because the story files are copied to a temporary folder it is not possible to include the original full story path.Instead only the file name is included.
If a recursive folder structure is used with the same story file names it can still be hard to find the problem file.