Update class Message to properly set attribute text (#5723)
Method `set` of class `Message` does not properly set attribute `text`. I know that it can be set as in `message.text = "foo"`, but method `set` should work as a unified API with method `get` for an intuitive API. Setting attribute `text` might be required for noisy text normalizers as a custom component, and calling `message.set('text', 'bar')` would not work just as in my case. I had to read the source code to make it work for me.