Fixes https://github.com/RasaHQ/rasa/issues/9949
New intent creation using `rasa interactive` was failing with server
returning 500. Upon closer inspection we found this was due to
prediction in unexpected intent policy; new intent is not in the domain
yet and this was raising an exception.

This is fixed by preventing predictions inside unexpected intent policy
for new intents.