Fix --augmentation parameter (#10181)
`--augmentation` was being mapped to a key not used by the TrainingTrackerProvider (augmentation instead of augmentation_factor), so when the component config (generated from the CLI args) was merged with the default config, the default value for `augmentation_factor` was never replaced. This fixes the mapping to make sure this doesn't happen any more when the parameter is specified.